+++ Public/Batch/Wait-Batch.ps1
@@ -105,7 +105,7 @@ function Wait-Batch {
         }
 
         try {
-            $PollCounter = 0
+            [uint]$PollCounter = 0
             $ProgressTitle = 'Waiting for completion...'
             do {
                 #Wait